Refactor dashboard components into separate files
