+++ rehearsal-room-frontend/src/components/Dashboard/index.jsx
@@ -1,71 +1,15 @@
-import { useEffect, useState } from 'react';
+
+// Custom Components
+import HostDashboardItems from './HostDashboardItems'
 
 // Material UI Components
 import { palette } from '@material-ui/system';
-import { makeStyles } from '@material-ui/core/styles'
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
-import Card from '@material-ui/core/Avatar';
-
-//AXIOS REQ GOES HERE
-const dummyBookings = [
-  {
-    id: 1,
-    user_id: 2,
-    requester_name: "Plimbus Perkins",
-    space_name: "The Big Cahoozey",
-    space_id: 1,
-    start_time: "05/10/2021 12:00:00",
-    end_time: "05/10/2021 2:00:00",
-    status: "pending",
-    usage_description: "I would like to throw a sick rave party for me and all my cool friends",
-    used_before: true,
-  },
-  {
-    id: 2,
-    user_id: 2,
-    requester_name: "Plimbus Perkins",
-    space_name: "The Rat Maze",
-    space_id: 1,
-    start_time: "05/10/2021 12:00:00",
-    end_time: "05/10/2021 2:00:00",
-    status: "rejected",
-    usage_description: "i'm gonna wreck it",
-    used_before: true,
-  },
-  {
-    id: 3,
-    user_id: 2,
-    requester_name: "Plimbus Perkins",
-    space_id: 3,
-    space_name: "The Flop Garden",
-    start_time: "05/10/2021 12:00:00",
-    end_time: "05/10/2021 2:00:00",
-    status: "confirmed",
-    usage_description: "i'm gonna wreck it",
-    used_before: true,
-  },
-  {
-    id: 4,
-    user_id: 2,
-    requester_name: "Jango Mango",
-    space_name: "The Flop Garden",
-    space_id: 2,
-    start_time: "05/10/2021 12:00:00",
-    end_time: "05/10/2021 2:00:00",
-    status: "confirmed",
-    usage_description: "i'm gonna wreck it",
-    used_before: true,
-  }
-]
 
 export default function Dashboard(props) {
   const { user } = props;
-  const [bookings, setBookings] = useState([])
-  useEffect(() => {
-    setBookings(dummyBookings)
-  })
 
   return (
     <Container maxWidth="lg">
@@ -83,35 +27,7 @@ export default function Dashboard(props) {
             <p><strong>{user.organization_name}</strong></p>
           </Paper>
         </Grid>
-        <Grid
-          container
-          direction="column"
-          justify="flex-start"
-          xs={9}
-          spacing={2}>
-          <Grid item>
-            <Paper>My Spaces
-            </Paper>
-          </Grid>
-          <Grid item>
-            <Paper>
-              Pending Booking Requests
-              <BookingList 
-                bookings={bookings}
-                bookingType="pending"
-              />
-            </Paper>
-          </Grid>
-          <Grid item>
-            <Paper>
-              Confirmed Bookings
-              <BookingList
-                bookings={bookings}
-                bookingType="confirmed"
-              />
-            </Paper>
-          </Grid>
-        </Grid>
+        { user.is_host && <HostDashboardItems />}
 
 
       </Grid>
@@ -119,35 +35,6 @@ export default function Dashboard(props) {
   )
 }
 
-function BookingList(props) {
-  const { bookings, bookingType } = props;
 
-  const bookingList =
-    bookings
-    .filter(booking => booking.status === bookingType)
-    .map(booking => (
-      <BookingListItem {...booking} />
-    ))
-  return (
-    <ul>
-      {bookingList}
-    </ul>
-  )
-}
 
-function BookingListItem(props) {
-  const shortenedUsageDesc = (desc) => {
-    if (desc.length > 40) {
-      return `${desc.slice(0, 40).trim()}...`
-    } else {
-      return desc
-    }
-  }
-  return (
-    <li>
-    <strong> From: </strong>{props.requester_name} |
-    <strong> Space: </strong>{props.space_name} |
-    <strong> Description: </strong>{shortenedUsageDesc(props.usage_description)} |
-    <strong> Time: </strong>{props.start_time} to {props.end_time}</li>
-  )
-}
+
